Update UI to support Marketing Actions Links
